Code refactoring part III
